+++ fastly/block_fastly_service_logging_logshuttle_test.go
@@ -38,7 +38,7 @@ func TestResourceFastlyFlattenLogshuttle(t *testing.T) {
 					"format":             "%h %l %u %t \"%r\" %>s %b %T",
 					"placement":          "none",
 					"response_condition": "always",
-					"format_version":     uint(2),
+					"format_version":     2,
 				},
 			},
 		},
@@ -98,10 +98,8 @@ func TestAccFastlyServiceVCL_logging_logshuttle_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceVCLExists("fastly_service_vcl.foo", &service),
 					testAccCheckFastlyServiceVCLLogshuttleAttributes(&service, []*gofastly.Logshuttle{&log1}, ServiceTypeVCL),
-					resource.TestCheckResourceAttr(
-						"fastly_service_vcl.foo", "name", name),
-					resource.TestCheckResourceAttr(
-						"fastly_service_vcl.foo", "logging_logshuttle.#", "1"),
+					resource.TestCheckResourceAttr("fastly_service_vcl.foo", "name", name),
+					resource.TestCheckResourceAttr("fastly_service_vcl.foo", "logging_logshuttle.#", "1"),
 				),
 			},
 
@@ -110,10 +108,8 @@ func TestAccFastlyServiceVCL_logging_logshuttle_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceVCLExists("fastly_service_vcl.foo", &service),
 					testAccCheckFastlyServiceVCLLogshuttleAttributes(&service, []*gofastly.Logshuttle{&log1AfterUpdate, &log2}, ServiceTypeVCL),
-					resource.TestCheckResourceAttr(
-						"fastly_service_vcl.foo", "name", name),
-					resource.TestCheckResourceAttr(
-						"fastly_service_vcl.foo", "logging_logshuttle.#", "2"),
+					resource.TestCheckResourceAttr("fastly_service_vcl.foo", "name", name),
+					resource.TestCheckResourceAttr("fastly_service_vcl.foo", "logging_logshuttle.#", "2"),
 				),
 			},
 		},
